requirements.txt
accelerate==0.24.1
ai2-olmo==0.2.5
aiofiles==23.2.1
aiohttp==3.8.4
aiosignal==1.3.1
almost-unique-id==0.0.3
evaluate==0.4.0
huggingface-hub==0.20.3
hydra-core==1.3.2
idna==3.4
ipython==8.15.0
jupyter==1.0.0
matplotlib==3.7.1
matplotlib-inline==0.1.6
numpy==1.25.0
pandas==2.0.3
seaborn==0.12.2
tabulate==0.9.0
torch==2.1.0
transformers==4.38.2
ruff
